style(dateFormatter.js): renamed files to closer match current naming convention
